+++ app/src/main/java/com/example/bookitup/MainActivity.java
@@ -61,8 +61,7 @@ public class MainActivity extends AppCompatActivity {
             // Passing each menu ID as a set of Ids because each
             // menu should be considered as top level destinations.
             mAppBarConfiguration = new AppBarConfiguration.Builder(
-                    R.id.nav_home, R.id.nav_books, R.id.nav_slideshow,
-                    R.id.nav_tools, R.id.nav_share, R.id.nav_send, R.id.nav_profile_edit)
+                    R.id.nav_home, R.id.nav_books, R.id.nav_order, R.id.nav_chat, R.id.nav_profile_edit)
                     .setDrawerLayout(drawer)
                     .build();
             NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
