+++ app/src/main/java/com/example/bookitup/ui/order/OrderHistoryFragment.java
@@ -1,4 +1,4 @@
-package com.example.bookitup.ui.tools;
+package com.example.bookitup.ui.order;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,17 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.bookitup.R;
 
-public class ToolsFragment extends Fragment {
+public class OrderHistoryFragment extends Fragment {
 
-    private ToolsViewModel toolsViewModel;
+    private OrderHistoryViewModel orderHistoryViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        toolsViewModel =
-                ViewModelProviders.of(this).get(ToolsViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_tools, container, false);
-        final TextView textView = root.findViewById(R.id.text_tools);
-        toolsViewModel.getText().observe(this, new Observer<String>() {
+        orderHistoryViewModel =
+                ViewModelProviders.of(this).get(OrderHistoryViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_order, container, false);
+        final TextView textView = root.findViewById(R.id.text_order);
+        orderHistoryViewModel.getText().observe(this, new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
