+++ app/src/main/java/com/example/bookitup/ui/order/OrderHistoryViewModel.java
@@ -1,14 +1,14 @@
-package com.example.bookitup.ui.slideshow;
+package com.example.bookitup.ui.order;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class SlideshowViewModel extends ViewModel {
+public class OrderHistoryViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public SlideshowViewModel() {
+    public OrderHistoryViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is slideshow fragment");
     }
