--- app/src/main/java/com/example/bookitup/ui/share/ShareViewModel.java
@@ -1,19 +1,0 @@
-package com.example.bookitup.ui.share;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class ShareViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public ShareViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is share fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
