+++ app/src/main/res/layout/fragment_chat.xml
