+++ app/src/main/res/values/strings.xml
@@ -6,17 +6,15 @@
     <string name="nav_header_subtitle">android.studio@android.com</string>
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
-    <string name="action_logout"><u>Log out</u></string>
+    <string name="action_logout">Log out</string>
     <string name="action_login_instead"><u>Already have a password? Login instead.</u></string>
     <string name="forgot_password"><u>Forgot Password?</u></string>
     <string name="action_edit_profile">Edit Profile</string>
 
     <string name="menu_home">Home</string>
     <string name="menu_books">My Books</string>
-    <string name="menu_slideshow">Order History</string>
-    <string name="menu_tools">Chat</string>
-    <string name="menu_share">Share</string>
-    <string name="menu_send">Send</string>
+    <string name="menu_order">Order History</string>
+    <string name="menu_chat">Chat</string>
     <string name="menu_profile">Profile</string>
 
     <string name="title_activity_edit_profile">EditProfileActivity</string>
